fix: Ctrl+C should work after exceptions (#332)
The added comment goes into a bit more detail here, but basically,
our shell evaluation function was under the wrong assumption that the
built-in eval code would always call its callback, even if evaluation
of the code threw an error. That does not turn out to be the case
for Node.js; instead, the REPL would handle the error internally
and not call its callback, leaving us with an async function that
never finishes.

This had the side effect that listeners for SIGINT would not be removed
(because from the cli-repl perspective, the inner eval would never
finish), leading to odd results when entering Ctrl+C after an earlier
statement had thrown an exception.

The solution here is hacky; we should probably take the time
at some point to find a better one.