MONGOSH-376 - Include metadata about data source in result
For results from calls on collections, add metadata about the source of
the returned data (database name, collection name, method, arguments)
so that consumers of the shell can figure out from what source
the result was generated, e.g. in order to understand how to modify
that original data.

This also adds a fake `Document` type to the `ShellResult` reported
for e.g. `findOne()`, so that it is possible to tell those apart
from other plain JavaScript objects, e.g. in order to be able to
know whether the result is a modifiable object in the database or not.