fix: Check if package exists while checking the access rights before publishing (#399)
In cases where new package was added to the repo and is published for the first time with the others
the access rights check can return a false-positive while checking for access rights: the package
is not on remote yet so read-write is not return. In addition to checking the access rights we will
additionally check if the package exists already to allow publishing new packages.