fix(service-provider-server): perform db caching MONGOSH-464 (#431)
We disabled caching the database objects on the driver level so that
we would be able to access the same database multiple times with
different options, and not get the same cached instance (with the
options from the first call) every time.

This has the major drawback that every time we do any operation that
uses such a database object, we create a new one. Due to the way
the driver creates those objects, they essentially cannot be gargage
collected for the lifetime of the `MongoClient` instance.
This is a memory leak that can crash mongosh.

To resolve this, we perform our own caching of the database instance
that also takes the options into account.