chore(test): download correct mongod version before running tests
Do this for the same reason that we install mlaunch during the
`pretest` script as well, namely, to account for the case in which
multiple test processes would otherwise attempt to download `mongod`
at the same time and then fail intermittently due to race conditions.

This should fix some more flakiness in recent CI runs.