chore(build): add way to write uploaded URL to file (#569)
Add a way to write the S3 URL for an uploaded package to a specified
location on disk. This is useful for MONGOSH-506, because it makes it
possible to store that file in a location that evergreen can understand
on its own (i.e. it doesn't need to use JS code to figure out where
the artifact is stored) and use it as a pointer to the actual artifact
URL.