fix(shell-api): use readPref for runCommand if explicitly set MONGOSH-508
If `readPreference` has been explicitly set, either through
a) the connection string or b) by using `setReadPref()`,
then pass it for `db.runCommand()` as well.

The reason we’re not always using it is that the driver will
report it back to use as `primary` by default, which would then
break `db.runCommand()` for almost all calls when we’re connected
to a secondary.

As part of this, update the default `Mongo` instance to also
report the connection string, like a programmatically created `Mongo`
instance would (and like it worked in the old shell).

Also addresses MONGOSH-557.

Refs: https://docs.google.com/document/d/10FTlhliWtg_XquUGbEmZQ7ccUiImMLlBT5Qlm2O1ZxI/edit?pli=1