fix(shell-api): use primaryPreferred for `show` commands
People don’t necessarily expect `show` to follow the same rules as
other shell calls, and expect it to work even when connected to a
secondary. Use `primaryPreferred` as the read preference for
`show db`/`show collections` to address this.
(This is a deviation from the old shell’s behavior.)