fix(cli-repl): disregard autocomplete eval for input stream MONGOSH-586 (#660)
This is the second half of MONGOSH-586.

The REPL’s built-in autocompleter uses `repl.eval()` to figure out
what JS object’s properties it would be autocompleting. These
`.eval()` calls don’t represent a state change as far as input
handling is concerned, so they should not affect the prompt and not
enable block-on-newline.