chore(cli-repl): add Segment API integration test MONGOSH-363
Add a test that verifies the actual API key bundled into the actual
release binary. In order to achieve this, add an internal test command
that performs a Segment API request and verifies that its callback
does not return an error.

Note that this is, currently, a test that with somewhat limited
meaningfullness, because the Segment API currently always gives
successful HTTP responses, regardless of the validitity of the API key.