fix(cli-repl): allow interrupting load()ed code MONGOSH-641
This was previously broken because the async repl implementation
adds an `process.on('SIGINT')` listener for the duration of the
*async* part of the evaluation. That listener would interfere with
breaking on sigint if another *synchronous* evaluation would be
started while it is present.

Fix this by not being lazy anymore and temporarily disabling
existing `SIGINT` event listeners on both `process` and `repl`
(and not just `repl`). In this case, the “inner” evaluation
also temporarily removes the “outer” `SIGINT` listener and
can go about as usual.