feat(async-rewriter2): introduce new async-rewriter alternative
See the package’s README.md for details on the why and how.
This is feature-flagged behind the `MONGOSH_ASYNC_REWRITER2`
env var and introduced as a separate package in this commit,
but if we do adopt this it’s going to enable us to remove
a lot of code *and* have a more flexible async-rewriting system.