fix(ci): work around CI Windows git failure MONGOSH-660
For some reason, Windows builds started failing recently.
The windows-64-vs2019-small image, the Node.js version,
the npm version, the git version, and the package-lock.json
all remained the same in that period, so it is unclear
what exactly caused this.

While a full investigation would likely be interesting,
that’s not easy with Evergreen’s current capabilities.

However, fixing this seems possible by explicitly adding
the git binaries (including libexec ones) to the PATH,
so let’s do that for now.