fix(async-rewriter2): fix interaction with domains MONGOSH-662
The fix I previously inserted for the “Node.js prints a warning
if we intentionally disregard a rejected Promise from the generated
inner async function” problem turned out to be problematic, because
it creates an infinite loop in the domain handler system.

This problem goes away with Node.js v14.15.2 (due to
https://github.com/nodejs/node/commit/0ad4f70db596934b3ff390cde7380ab8bf5480f5),
but since we also support Node.js-v12.x environments
(and we would want to be resilient against changes to the
platform internals anyway), it makes sense to aim for a better
solution here.

This patch makes sure that the generated inner async function only
returns a meaningful value if its return value is actually used,
i.e. if we’re encountering an async-evaluation case where the
generated outer function actually just forwards the returned
Promise.