chore(cli-repl): (hopefully) fix flaky mongocryptd test (#754)
Example failure:

    [2021/04/01 09:26:24.908]   1) MongocryptdManager
    [2021/04/01 09:26:24.908]        with network testing
    [2021/04/01 09:26:24.908]          performs keepalive pings:
    [2021/04/01 09:26:24.908]      SyntaxError: Unexpected end of JSON input
    [2021/04/01 09:26:24.908]       at JSON.parse (<anonymous>)
    [2021/04/01 09:26:24.908]       at Context.<anonymous> (src/mongocryptd-manager.spec.ts:184:19)

Presumably this is happening because the connections aren’t established
fast enough when the host is under load, e.g. because multiple tests
are running at the same time. Using the `eventually` helper would be
the right thing to do here anyway, because it allows for a longer
timeout while at the same time finishing faster under normal
circumstances.