chore(shell-evaluator): do not log full rewritten code (#786)
With the new async rewriter, the rewritten code is unfortunately quite
a bit larger in size than the input, so logging both input and output
leads to large log files. However, since the new rewriter is stateless,
we also don’t need to worry about the output anymore, and can just
skip logging it.