chore(cli-repl): fix flaky test and improve errors
The Node.js REPL produces errors like this when a module is not found:

```js
Uncaught Error: Cannot find module 'a'
Require stack:
- <repl>
[ ... ]
```

Our test helpers interpreted the `- <repl>` string as a prompt, which
it is not.

Fix that by requiring the prompt not to contain a matching
angle bracket, and provide more precise error messages from our
test helpers in general to help debug future problems.