fix(shell-api): fix prompt for Atlas-proxy-style mongos MONGOSH-680
When connected to a server that we know is an Atlas server but
announces itself as a mongos, print “atlas proxy” in the prompt
rather than “mongos” to avoid confusion.

As a drive-by fix, make sure to use the service provider of the
current database, and not necessarily the initial one, as new
connections can be assigned to `db`.