feat(cli-repl): aggregate .editor multiline input MONGOSH-518
The changes for this feature are fairly self-contained, however,
.editor support relies a lot more strongly on automated semicolon
insertion (ASI) than the previous multi-line input handling,
so the bigger part of the changes here is getting the semantics for
that right. (This would also have affected the original multi-line
aggregating functionality, e.g. `() => { 1\n2 }` would have resulted
in a broken `() => { 1 2 }` history entry.)