chore(shell-api): mention --tls for connect errors MONGOSH-814
Mention that `tls=` needs to be explicitly specified when passed
to `new Mongo()`/`connect()`, because unlike in the legacy shell,
that setting is not inherited. This is only printed for connection
attempts from a shell that was started with TLS enabled using
connection strings that do not contain a TLS flag which encounter
ServerSelectionTimeout errors.