chore(ci): test s390x builds on RHEL 8.0 MONGOSH-768 (#980)
The platform support matrix for the 5.0 server now indicates that
RHEL 8 will be supported in addition to, not instead of, RHEL 7
on the zSeries/s390x platform.

This is the optimistic approach, where we build the same executable
for both targets. If this fails in CI, we may need to start
building separate binaries.