fix(service-provider-server): fix change stream .next() interrupt MONGOSH-888
When a user interrupts running code while a change stream `.next()`
call is running, that closes the network connection. However,
network errors are considered retryable by the driver (in line
with the spec); this makes the change stream cursor `.next()`
operation attempt to re-try itself, which fails because the
MongoClient is already unusable at that point.

In order to address this, we call the driver-internal
`Connection.handleIssue()` method (which should be okay, since the
relevant code here is concerned only with driver internals anyway)
with a non-network-error state, which still closes ongoing operations.