fix(cli-repl): return original evaluation value to REPL MONGOSH-902
Instead of returning the ShellResult interpretation of the evaluation
result back to the Node.js REPL, return the original raw value and
keep the ShellResult stored separately (which still needs to happen
during `eval`, as the formatting itself is a synchronous operation).

This ensures that `_` will always refer to the actual value of the
last evaluation result.