chore(shell-api): make coll.find() an async function
This is in preparation for MONGOSH-537, where it will make things a bit
easier if we have the liberty to make async calls inside all methods
that perform interactions with the database (i.e. fetching
`_baseOptions` can be an async operation).

Also tighten up some types in the tests along with this.