chore(cli-repl): move arg parser warning printing to run.ts (#1089)
This is just a small bit of cleanup. The arg parser should not
be printing warnings on its own, without knowing where those end up;
specifying the output stream is the top level program’s responsibility.

This also makes running the arg parser test suite a bit nicer, because
now the test output is not mixed with these warnings.