fix: do babel transformations for script, not module MONGOSH-986
Tell babel that we are targeting script mode, not ES module mode.
(Note that the babel docs indicate that `script` should be the
default – it seems that that is not entirely accurate.)

This fixes the linked bug because unlike ES modules, script mode
does not automatically also enable strict mode.

This also fixes a minor bug in the wrap-as-iife babel plugin,
where if the input was a single string literal, escaped characters
would have remained escaped (e.g., entering `'\x41'` in mongosh
would have returned `\x41`, not 'A'), by keeping the existing
information about the raw source instead of generating an
entirely new string literal in that case.