fix(shell-api): .watch() signature and behavior fixes MONGOSH-976
Modify the `.watch()` signature so that it “officially” allows
omitting the first argument.

As a drive-by fix, stop applying `.tryNext()` for cases in which
the change stream cursor starts from a specific point in time.
(Unfortunately, `.hasNext()` is not an option here, unlike
discussed with the Node.js driver team, because it is blocking
like `.next()` and unlike `.tryNext()`.)