fix(async-rewriter): do not load browserlist config MONGOSH-1085
We should tell babel not to try to load the browerlist config from
disk, especially since we are not interested in it anyway.

A proper regression test seems hard to write, since it would
essentially consist of making sure that mongosh does not access
files from its current working directory when it starts up.