fix(ci): fix Windows packaging URL upload
In c9151ce70b79, I removed a layer of indirection in
`package-and-upload-artifact.sh`, which previously wrote most
of its content into another bash script and then executed it.

However, this also mean that I should have adjusted the escaping
to be one layer less deep, which I forgot and which causes
CI to be red on Windows right now.