fix(node-runtime-worker-thread): externalize system-ca and deps COMPASS-4105
In order to be usable in Compass, system-ca and its addon dependencies
need to be externalized in the webpack configuration.

Theoretically, it might be enough to only externalize the addons;
however, since system-ca prefers to use worker threads for loading
certificates asynchronously on macOS and Windows, and we need to
externalize *something*, it makes sense to do it on the system-ca
level.