chore(deps): bump Node.js to latest versions MONGOSH-1163 (#1240)
* chore(deps): bump Node.js to latest versions MONGOSH-1163

* fixup: try building linux-arm64 on Amazon2 instead of Ubuntu 18.04

Attempt to address a glibc versioning mismatch by building on
Amazon 2 instead of Ubuntu 18.04. The version bump from Node.js
v14.18.3 to v14.19.0 includes an ICU update, which introduces
a call to `expf`, which uses a symbol tagged with GLIBC_2.27
when compiled on Ubuntu 18.04; that glibc version is not available
on Amazon Linux 2.

If possible, building on Amazon Linux 2 instead of Ubuntu 18.04
should be the best way of resolving this issue.