fix: explicitly call ObjectId.toHexString() in internals MONGOSH-1097
Call `.toHexString()` when we convert an `ObjectId` to a string
for internal usage.

This ensures that changes to `ObjectId.prototyp.toString()` from
user code will not affect mongosh internals.

Unfortunately, it is not easily possible to add a regression test
for the original ticket here, since it depends specifically on
the way in which homebrew (or npx) install mongosh, where dependencies
of the different packages in the monorepo are deduplicated during
installation (which makes `service-provider-server` and
`service-provider-core` share a `bson` dependency rather than
using separate ones).