feat(cli-repl): add --json CLI flag MONGOSH-1249
Add a `--json` CLI flag with possible `--json=canonical` and
`--json=relaxed` flags, the default being `canonical`.
When used, mongosh will print Extended JSON as output. This
flag can only be used in conjunction with `--eval`, and is
mutually exclusive with `--shell` and providing files on
the command line.

Errors and regular output are not distinguished based on
the output alone, only in the exit code of mongosh.
Repeated errors while trying to serialize the output to
Extended JSON will not be serialized.