chore: add driver as explicit dependency where it is a peer dep
Add the driver package as an explicit `devDependencies` entry
for packages where a dependency uses it as a peer dependency.

This is not strictly necessary with modern npm versions,
and hopefully goes away once we move to a more Compass-style
monorepo where dependencies are hoisted, but for now, this is
convenient because it makes it quite a bit easier to update
the driver, and hopefully also paves the way for fixing the
(currently broken) driver integration tests.

(Previously, updating the driver often involved adding the
driver as a dependency to these packages, running bootstrap
to update the package-lock.json, then removing it from the
package.json again, and then running bootstrap again, since
otherwise driver types would start mismatching the driver types
from other packages.)