fix(shell-api): allow pipeline-style bulk update operations MONGOSH-1281
Fix cloning of the update documents to allow arrays as well as
objects, so that users can use aggregation pipelines for bulk updates.

Also, remove the buggy `hint` and `arrayFilters` implementation
that was storing extra properties in the wrong place on the outgoing
documents, and instead fully rely on driver helpers to implement
these.