fix(cli-repl): do not enable telemetry unless user id is persisted MONGOSH-1320
Do not enable telemetry unless a new config file has been written
or an existing one read.

The alternative would be to send telemetry even when user ids cannot
be persisted; this would make it appear as if a new user connected
every time mongosh runs, which is probably not what we want.