chore(deps): bump `mongodb-client-encryption` to 2.4.0
Also add `@aws-sdk/credential-prociders` as an explicit dependency.

The `mongodb-client-encryption` package can use this as an optional
peer dependency, and the `mongodb` package currently depends on this
as a peer dependency but will switch to making it an optional peer
dependency as well in the v5.0.0 release.

Installing it as an explicit dependency ensures its presence regardless
of optionality (instead of just relying on the package-lock.json file
for keeping it available).