fix(cli-repl): lazy-load dependencies in entry point MONGOSH-1336
Lazy-load most dependencies in the main entry point so as not to
have to evaluate their code when running mongosh in very simple
cases like `--version` and `--build-info`, particularly because
the cloud team needs these to be relatively fast.

This results in a ~75% performance improvement for `mongosh --version`,
although it is likely that that difference will be reduced later
once/if we enable snapshot support.