fix(shell-api): fix broken tsToSeconds conversion MONGOSH-1428 (#1433)
`tsToSeconds()` has not been working as intended since it
was first written. BSON `Timestamp` objects never had
direct `t` or `i` properties, so the `if` condition here
would never be met. Instead, if the function received
a `Timestamp` object, it would implicitly convert it to a
number (via the `Long` class’s `.toString()` conversion method).
This happens to work out because `x/(2**32)` corresponds
roughly to the time part for a `Timestamp` object `t` when
interpreted as a `Long`.

I am not sure why exactly this breaks in the way it does
described in the ticket (where this function seems to
return `NaN`), but making this function work as intended
should address the issue.