fix(cli-repl): invert and fix passwordless-auth-mechanism check
Currently, if a username is specified but no password, mongosh
will prompt for a password before connecting.

Kerberos/GSSAPI was made an exception, since it rarely makes use
of a password. However, there are other mechanisms where a username
can, and often will, be provided without a password (AWS, OIDC,
and technically even X.509).

This commit prevents mongosh from prompting for a password for those
mechanisms, and inverts the check to explicitly list the mechanisms
that *do* require a password (the list of auth mechanisms is unlikely
to change soon, but if it does, it likely wouldn’t include a new
password-based auth mechanism, so avoiding false positive return values
seems like a better choice than avoiding false negative return values).