fix(build): keep using 6.0.x RHEL7.2 crypt_shared library on s390x
The server team just release 7.0.0, which does not provide a
crypt_shared library for RHEL7 on s390x. To keep things as they
are currently working, we continue to use the 6.0.x shared library
instead. We may want to switch to the RHEL8 7.x-server library
at some point in the future when there are new QE features that
require an updated shared library.