fix(ci): use npm@9.x
Previously, our CI on Node.js 20 was failing silently (e.g.
the bug fixed by https://github.com/mongodb-js/mongosh/pull/1591
did make mocha exit with a non-zero exit code, but not our CI).

This is being caused by an incompatibility between npm@8.x and
Node.js 20 (which could reasonably be argued to be a Node.js bug,
the same one that the PR mentioned above accounts for). Upgrading
to npm@9.x should fix this because it includes a fix in npm
as well (https://github.com/npm/cli/issues/6399).