fix(shell-api): use public BulkWriteResult props MONGOSH-1507 COMPASS-7023
This accounts for removal of these legacy getters in the 6.x driver.

The only catch here is that `.result.upserted` is of a different type
than `.upsertedIds`, but since the type of the latter matches what our
own `BulkWriteResult` class expects (and also matches `.insertedIds`),
this is probably closer to a bug fix than proper breakage.