fix(build): fix list of files expected to be unchanged for releases
Before publishing to npm, lerna ensures that the working tree is
in a clean state (no unchanged or uncommitted files).

Our logic to make `git` assume that files were unchanged was
broken for a few files, in particular the top-level package.json
and the `.nvm` directory that we now place under the `.evergreen/`
folder in the repository rather than a temporary path.