chore(ci): add qemu-based regression test for architecture minima (#1835)
To fix MONGOSH-1604, we temporarily disabled startup snapshots in
cd6e6af174b. As we are now starting to look into re-enabling this
feature again soon, we should have a regression test in place if
it’s possible.

This commits adds a docker imagine within which we install qemu-user
to emulate the minimum requirements for our supported architectures
on arm64 and x64.