fix(build): fix hash file parsing
In 2da3c3159545e, we moved file hash computation for our produced artifacts
from the JSON feed generation step to the newly introduced "download artifacts"
step of the `release_publish` tasks (because we need to download artifacts
before we can submit them to the papertrail service). We used the standard-ish
SHASUMS format (output of `sha256sum`/input of `sha256sum -c`) for that, but
failed to parse it properly when consuming it.

This fixes that and extends the download center config tests to also perform
the full "download artifacts" step as part of the test setup, to ensure that
the two parts work together properly.