fix(autocomplete): handle missing `buildInfo` well MONGOSH-1792
Recent server source changes put the `buildInfo` command behind
authentication. We previously assumed that this information would
always be available (and, as a corollary, that if the information
from `buildInfo` was absent, topology information would also be absent),
and should stop assuming this.

Long-term, it is probably desirable to rely on `buildInfo` as little
as possible – for example, most of our version checks for autocompletion
should probably be wire version checks instead of server release version
checks.

This should address some failures occuring in the latest-alpha CI jobs.