chore(release): use a different release process when publishing auxiliary packages (#2319)
When doing a mongosh release, we have different checks in place to ensure all is setup as expected. We also run additional tasks to publish in places like homebrew.  At the moment we just need publish to npm when publishing auxiliary packages. This commit more clearly separates those two cases and allows a more permissive and simpler release setup for auxiliary packages.