refactor(ci): use more object-oriented structure during publishing and testing
This ended up being a larger refactor than intended but it was getting increasingly difficult to add new helpers and test them with the current setup. Now, instead of passing all helpers as spies, they exist as properties on "Publisher" objects which makes them easier to stub them with sinon.