fix(ci): use the correct commit message when bumping packages
bump-packages from monorepo-tools is looking for chore(ci): bump packages as the start of the commit to calculate the semver bump. The problem is that our commit message doesn't start with this - instead it's chore(release): bump auxiliary packages. This means that the next time we run it, we go back to the start of time and sure enough, we find some breaking: ... commit message and decide we need to bump major. Luckily there's a LAST_BUMP_COMMIT_MESSAGE we can use to set this.