fix(cli-repl): support bracketed paste on Windows MONGOSH-1999
Split out into a separate piece because this adds a patch to libuv
that has non-trivial downsides, as explained in the referenced PR.
I'd assume that mitigating this through a `process.on('exit')` handler
is sufficient and that we don't need to plan specifically for e.g.
hard crashes of mongosh here, but I didn't want to merge this together
with the original work in 4232a9847cd5d2486 but rather have it stand
on its own (including for the simple reason that it's easier to revert
if we decide that we have to).