fix(tests): minimize flakiness of the disableLogging test
There seems to be flakiness in some of the recently introduced disableLogging tests, likely because the needed log doesn't appear in the file quick enough for us to capture it. This adds a `eventually` to the tests where we check for a new log entry after inputting a command so we can retry at a later time and minimize flakiness. This makes it consistent with other log file tests.